Matts Suggestion Update
Update made based on Matts suggestion:
Looks good. I think we could shorten the Declaration of Ethical Use to something like:

I declare that AI was used in this project as outlined above. The following ethical issues were considered when developing the project: (Delete or add your own as necessary)

* Gender or sexual identity bias
* Racial bias
* Age bias
* Accessibility